finish work build (fix bug and refactor logic)

+++ src/main/java/dto/ExchangeRateDTO.java
@@ -2,17 +2,19 @@ package dto;
 
 import models.Currency;
 
+import java.math.BigDecimal;
+
 public class ExchangeRateDTO {
-    private Integer id;
+    private int id;
     private Currency baseCurrency;
     private Currency targetCurrency;
-    private Double rate;
+    private BigDecimal rate;
 
-    public void setId(Integer id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public void setRate(Double rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
 
@@ -28,7 +30,7 @@ public class ExchangeRateDTO {
         return id;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
